requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
strawberry-graphql[fastapi]==0.235.1
databases[asyncpg]==0.9.0
sqlalchemy==2.0.29
aio-pika==9.4.1
redis==5.0.4
httpx==0.27.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.7.4
psycopg2-binary==2.9.9